python-ai/requirements.txt
# ThinkBank AI Service Requirements
# Optimized for RTX 4070 8GB VRAM

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# LLM Inference (vLLM)
vllm>=0.6.0

# LangChain Orchestration
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# Embedding Models
sentence-transformers>=3.0.0
transformers>=4.45.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4
sqlalchemy>=2.0.0

# Object Storage
minio>=7.2.0

# Task Queue
redis>=5.0.0

# Document Parsing
docling>=2.0.0
# or marker-pdf for alternative

# Image Processing
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
loguru>=0.7.0
